Rename package 'debug' to 'logger'
