+++ src/main/java/com/ruralhousejsf/model/availability/QueryAvailabilityBean.java
@@ -3,6 +3,7 @@ package com.ruralhousejsf.model.availability;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
+
 import javax.faces.application.FacesMessage;
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.SessionScoped;
